+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/config/domain/DefaultProvisioningBreakConfiguration.java
@@ -1,4 +1,4 @@
-package eu.bcvsolutions.idm.acc.domain;
+package eu.bcvsolutions.idm.acc.config.domain;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 
+import eu.bcvsolutions.idm.acc.domain.ProvisioningEventType;
 import eu.bcvsolutions.idm.core.api.config.domain.AbstractConfiguration;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmRoleDto;