+++ CHANGELOG.md
@@ -16,6 +16,12 @@ All notable changes to this project will be documented in this file.
 ##### Provisioning
 
 - After add, remove or update ``IdmContractGuaranteeDto`` is execute provisioning for identity that own this contract.
+- It's possible to send additional attributes, when password is changed (e.g. password expiration in extended attribute). New flag ``sendOnPasswordChange`` was added to system attribute mapping - attribute with this flag checked will be send together with changed password to provisioning. Two ways for provisioning additional attributes are implemented:
+  - send additional attributes together with new password in one provisioning operation
+  - send additional attributes after password is changed in another provisioning operation
+  - two ways are be configurable by application configuration ``idm.sec.acc.provisioning.supportSendPasswordAttributes`` (is effective for all target systems):
+    - ``true``: additional password attributes will be send in one provisioning operation together with password
+    - ``false``: additional password attributes will be send in new provisioning operation, after password change operation
 
 ### Changed